refactor: sectionService 삭제
