+++ src/main/java/subway/line/LineController.java
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import subway.line.dto.LineCreateDto;
 import subway.line.dto.LineResponseDto;
-import subway.section.SectionService;
 import subway.section.dto.SectionCreateDto;
 import subway.station.dto.StationResponseDto;
 
@@ -21,11 +20,9 @@ import subway.station.dto.StationResponseDto;
 public class LineController {
 
     private final LineService lineService;
-    private final SectionService sectionService;
 
-    public LineController(final LineService lineService, final SectionService sectionService) {
+    public LineController(final LineService lineService) {
         this.lineService = lineService;
-        this.sectionService = sectionService;
     }
 
     @GetMapping
@@ -35,24 +32,24 @@ public class LineController {
 
     @GetMapping("/{lineId}")
     public ResponseEntity<List<StationResponseDto>> getStation(@PathVariable final Long lineId) {
-        return ResponseEntity.ok(sectionService.findSortedStations(lineId));
+        return ResponseEntity.ok(lineService.findSortedStations(lineId));
     }
 
     @PostMapping
     public ResponseEntity<Void> create(@RequestBody LineCreateDto lineCreateDto) {
-        final Long lineId = lineService.create(lineCreateDto);
+        final Long lineId = lineService.createLine(lineCreateDto);
         return ResponseEntity.created(URI.create("/lines/" + lineId)).build();
     }
 
     @PostMapping("/stations")
     public ResponseEntity<Void> enrollStation(@RequestBody final SectionCreateDto sectionCreateDto) {
-        final Long lineId = sectionService.addSection(sectionCreateDto);
+        final Long lineId = lineService.addSection(sectionCreateDto);
         return ResponseEntity.created(URI.create("/lines/" + lineId)).build();
     }
 
     @DeleteMapping("/{lineId}/stations/{stationId}")
     public ResponseEntity<Void> disEnrollStation(@PathVariable final Long lineId, @PathVariable final Long stationId) {
-        sectionService.removeStationBy(lineId, stationId);
+        lineService.removeStationBy(lineId, stationId);
         return ResponseEntity.noContent().build();
     }
 }
